// libs/features/occupational-hazard/src/lib/shared/models/auditor-filter-params.ts
/**
 * Copyright GOSI. All Rights Reserved.
 * This software is the proprietary information of GOSI.
 * Use is subject to license terms.
 */

export class AuditorFilterParams {
  maxAllowances: number = undefined;
  maxNewAllowances: number = undefined;
  minAllowances: number = undefined;
  minNewAllowances: number = undefined;
  ohType: string[] = undefined;
}