// libs/features/occupational-hazard/src/lib/shared/models/allowance-list.ts
/**
 * Copyright GOSI. All Rights Reserved.
 * This software is the proprietary information of GOSI.
 * Use is subject to license terms.
 */
import { AllowanceSummary } from './allowance-summary';
import { GosiCalendar } from '@gosi-ui/core';

export class AllowanceList {
  auditCases: AllowanceSummary[];
  batchMonth: GosiCalendar = new GosiCalendar();
  tpaId: number;
  tpaName: string;
}